moved friendRequests utils
