+++ src/pages/api/users/profile/friends/interact-request.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { verifyUser } from "../../../../../utils/auth";
-import { deleteFriendRequest, getFriendRequest } from "../../../../../utils/friendRequests";
+import { deleteFriendRequest, getFriendRequest } from "../../../../../database/utils/friendRequests";
 import { addFriend } from "../../../../../utils/friends";
 import { getUser } from "../../../../../utils/users";
