+++ src/pages/api/users/profile/friends/requests-received.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { getAuthFromApi } from "../../../../../utils/auth";
-import { getFriendRequestsFromUser, getFriendRequestsToUser, getToOfFriendRequestsFromUser } from "../../../../../utils/friendRequests";
+import { getFriendRequestsFromUser, getFriendRequestsToUser, getToOfFriendRequestsFromUser } from "../../../../../database/utils/friendRequests";
 import { filterUser, getUser } from "../../../../../utils/users";
 
 export default async function FriendRequestsReceived(req:NextApiRequest, res:NextApiResponse) {
