+++ src/pages/api/users/profile/friends/requests-sent.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { getAuthFromApi } from "../../../../../utils/auth";
-import { getFriendRequestsFromUser, getToOfFriendRequestsFromUser } from "../../../../../utils/friendRequests";
+import { getFriendRequestsFromUser, getToOfFriendRequestsFromUser } from "../../../../../database/utils/friendRequests";
 import { filterUser, getUser } from "../../../../../utils/users";
 
 export default async function FriendRequestsSend(req:NextApiRequest, res:NextApiResponse) {
